requirements.txt
einops
kornia
natsort
matplotlib
numpy
omegaconf
openpyxl
pandas
pillow
psychopy
pylsl
pyqt
pyqtgraph
pytorch-lightning
pywavelets 
requests
scikit-image
scipy 
seaborn 
sklearn
timm
torch
torchmetrics 
torchvision 
tqdm 
transformers
typeguard==2.13.3
wandb